[add] int page total

+++ MyDAL/AdoNet/DataSourceAsync.cs
@@ -144,7 +144,7 @@ namespace HPC.DAL.AdoNet
                 using (var cmdC = SettingCommand(ci, Conn, ci.Parameter.ParamReader))
                 {
                     var obj = await cmdC.ExecuteScalarAsync();
-                    result.TotalCount = DC.GH.ConvertT<long>(obj);
+                    result.TotalCount = DC.GH.ConvertT<int>(obj);
                 }
                 ci = new CommandInfo(SqlTwo, Parameter);
                 using (var cmdD = SettingCommand(ci, Conn, ci.Parameter.ParamReader))
